dropoutwatch-app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
twilio==8.10.0
python-dotenv==1.0.0
WTForms==3.1.1
email-validator==2.1.0
